pyproject.toml
[project]
name = "alarmfw-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = ["httpx", "pytest", "pytest-asyncio"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends.legacy:build"
